Simplify for creating a cluster on premisses with ACM
Removed unneccesary parts of the project.

The ansible playbook has been modified and simplified to setup haproxy and DNS.  The http and openshift installer parts have been removed

The DNS configuration has been changed to apply just to the particular cluster subdomain and not the whole dynamic.openctlc.com, so that the other cluster DNS domain can also be resolved. This way the names for the cluster subdomain are resolved by the local DNS server, and any other DNS names are sent to the higher level DNS server.

The haproxy configuration has been changed to listen on the IP for the api and *.apps DNS records.  The configuration is intended for a 3 compact cluster with only masters, no workers.

Signed-off-by: Jose Jerez <tale.toul@gmail.com>